Fix telegram_id to BigInteger and add auto migration

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, Boolean, ForeignKey, Float
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DateTime, Boolean, ForeignKey, Float
 from sqlalchemy.orm import declarative_base, relationship
 from datetime import datetime
 
@@ -6,15 +6,15 @@ Base = declarative_base()
 
 class User(Base):
     __tablename__ = "users"
-    
+
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True)
+    telegram_id = Column(BigInteger, unique=True, index=True)
     username = Column(String(255))
     first_name = Column(String(255))
     last_name = Column(String(255))
     registered_at = Column(DateTime, default=datetime.utcnow)
     is_active = Column(Boolean, default=True)
-    
+
     # Relationship to partner profile
     partner_profile = relationship("PartnerProfile", back_populates="user", uselist=False, cascade="all, delete-orphan")
     service_requests = relationship("ServiceRequest", back_populates="user", cascade="all, delete-orphan")
